refactor: add scripts for reducer
